# Faster categorical column names selection