Change slow and redundant dataframe query by select_dtypes into a dataframe.dtypes list comprehension

```
import pandas as pd
import random
temp_df = pd.DataFrame({str(a):[random.random() for _ in range(100)] for a in range(200)})
temp_df[["100", "121", "115"]] = temp_df[["100", "121", "115"]].astype("category")
```

`%timeit temp_df.select_dtypes("category").columns`->`23.1 ms ± 1.04 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)`

VS

`%timeit [col for col, t in zip(temp_df.columns, temp_df.dtypes) if isinstance(t, CategoricalDtype)]` ->`167 µs ± 16.4 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)`